offline-llm-app/requirements.txt
llama-cpp-python
flask
pyinstaller
